pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["Subeom Jeong <92135821+MoebusChris@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
django-filter = "^23.5"
django-split-settings = "^1.2.0"
pyyaml = "^6.0.1"
boto3 = "^1.34.47"
pillow = "^10.2.0"
channels = "^4.0.0"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
django-storages = "^1.14.2"
pydantic = "^2.6.3"
pynacl = "^1.5.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pre-commit = "^3.6.2"
colorlog = "^6.8.2"
pytest = "^8.1.1"

[tool.isort]
line_length = 119
multi_line_output = 5

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = "google"
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
